requirements.txt
Django==3.1.7
django-mathfilters==1.0.0
python-dotenv~=0.17.0
django-choices~=1.7.1

Pillow==8.1.2

gunicorn==20.0.4

pre-commit==2.11.1

mccabe==0.6.1

psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1

pytest-django~=4.4.0
pytest-xdist~=2.4.0

django-3-jet~=1.0.8
django-static-fontawesome~=5.14.0.0

djangorestframework~=3.12.0
djangorestframework-simplejwt~=4.7.2
drf-yasg~=1.20.0

graphene~=2.1.8
graphene-django~=2.15.0

celery~=5.0.0
flower==1.0.0
redis~=3.5.3
django-redis~=4.12.1
